Better naming in the sample app
