+++ FountainKitDemo/AppDelegate.swift
@@ -3,7 +3,6 @@
 //  FountainKitDemo
 //
 //  Created by Cesar Tardaguila on 12/3/2016.
-//  Copyright © 2016 Cesar Tardaguila. All rights reserved.
 //
 
 import UIKit
@@ -17,7 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         self.window = UIWindow(frame: UIScreen.mainScreen().bounds)
         
-        self.window!.rootViewController = ViewController()
+        self.window!.rootViewController = MoviesViewController()
         
         self.window!.makeKeyAndVisible()
         return true
